+++ src/app/blog/blog.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { Observable } from 'rxjs';
-import { IBlog } from '../shared/models/blog.model';
+import { IBlog } from './blog.model';
 import { environment } from '../../environments/environment';
 
 @Injectable({
