+++ src/app/profile/profile.reducer.ts
@@ -1,7 +1,7 @@
 import { createEntityAdapter, EntityAdapter, EntityState } from '@ngrx/entity';
 import { createReducer, on } from '@ngrx/store';
 import * as ProfileActions from './profile.action';
-import { IProfile } from '../shared/models/profile.model';
+import { IProfile } from './profile.model';
 import { ViewStatus } from '../shared/constants/status.constant';
 import { equals, isNil } from '../shared/utils/ramda-functions.util';
