+++ src/app/shared/ui/map-icon/marker-icon.resource.ts
